OutfitManager.cs
