+++ Modules/Camouflague.cs
@@ -1,4 +1,5 @@
 using AmongUs.Data;
+using TOHE.Modules;
 using TOHE.Roles.Impostor;
 using TOHE.Roles.Neutral;
 
@@ -209,40 +210,6 @@ public static class Camouflage
         if (newOutfit.Compare(target.Data.DefaultOutfit)) return;
 
         Logger.Info($"playerId {target.PlayerId} newOutfit={newOutfit.GetString().RemoveHtmlTags()}", "RpcSetSkin");
-
-        // Start to set Outfit
-        var sender = CustomRpcSender.Create(name: $"Camouflage.RpcSetSkin({target.Data.PlayerName})");
-
-        target.SetColor(newOutfit.ColorId);
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetColor)
-            .Write(target.Data.NetId)
-            .Write((byte)newOutfit.ColorId)
-            .EndRpc();
-
-        target.SetHat(newOutfit.HatId, newOutfit.ColorId);
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetHatStr)
-            .Write(newOutfit.HatId)
-            .Write(target.GetNextRpcSequenceId(RpcCalls.SetHatStr))
-            .EndRpc();
-
-        target.SetSkin(newOutfit.SkinId, newOutfit.ColorId);
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetSkinStr)
-            .Write(newOutfit.SkinId)
-            .Write(target.GetNextRpcSequenceId(RpcCalls.SetSkinStr))
-            .EndRpc();
-
-        target.SetVisor(newOutfit.VisorId, newOutfit.ColorId);
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetVisorStr)
-            .Write(newOutfit.VisorId)
-            .Write(target.GetNextRpcSequenceId(RpcCalls.SetVisorStr))
-            .EndRpc();
-
-        target.SetPet(newOutfit.PetId);
-        sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetPetStr)
-            .Write(newOutfit.PetId)
-            .Write(target.GetNextRpcSequenceId(RpcCalls.SetPetStr))
-            .EndRpc();
-
-        sender.SendMessage();
+        target.SetNewOutfit(newOutfit, setName: false, setNamePlate: false);
     }
 }
