+++ Roles/Neutral/Doppelganger.cs
@@ -1,3 +1,4 @@
+using TOHE.Modules;
 using TOHE.Roles.Core;
 using TOHE.Roles.Impostor;
 using UnityEngine;
@@ -49,64 +50,6 @@ internal class Doppelganger : RoleBase
 
     public static bool CheckDoppelVictim(byte playerId) => DoppelVictim.ContainsKey(playerId);
 
-    private static void RpcChangeSkin(PlayerControl pc, NetworkedPlayerInfo.PlayerOutfit newOutfit, uint level)
-    {
-        var sender = CustomRpcSender.Create(name: $"Doppelganger.RpcChangeSkin({pc.Data.PlayerName})");
-
-        pc.SetName(newOutfit.PlayerName);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetName)
-            .Write(pc.Data.NetId)
-            .Write(newOutfit.PlayerName)
-        .EndRpc();
-
-        Main.AllPlayerNames[pc.PlayerId] = newOutfit.PlayerName;
-
-        pc.SetColor(newOutfit.ColorId);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetColor)
-            .Write(pc.Data.NetId)
-            .Write((byte)newOutfit.ColorId)
-        .EndRpc();
-
-        pc.SetHat(newOutfit.HatId, newOutfit.ColorId);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetHatStr)
-            .Write(newOutfit.HatId)
-            .Write(pc.GetNextRpcSequenceId(RpcCalls.SetHatStr))
-        .EndRpc();
-
-        pc.SetSkin(newOutfit.SkinId, newOutfit.ColorId);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetSkinStr)
-            .Write(newOutfit.SkinId)
-            .Write(pc.GetNextRpcSequenceId(RpcCalls.SetSkinStr))
-        .EndRpc();
-
-        pc.SetVisor(newOutfit.VisorId, newOutfit.ColorId);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetVisorStr)
-            .Write(newOutfit.VisorId)
-            .Write(pc.GetNextRpcSequenceId(RpcCalls.SetVisorStr))
-        .EndRpc();
-
-        pc.SetPet(newOutfit.PetId);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetPetStr)
-            .Write(newOutfit.PetId)
-            .Write(pc.GetNextRpcSequenceId(RpcCalls.SetPetStr))
-            .EndRpc();
-
-        pc.SetNamePlate(newOutfit.NamePlateId);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetNamePlateStr)
-            .Write(newOutfit.NamePlateId)
-            .Write(pc.GetNextRpcSequenceId(RpcCalls.SetNamePlateStr))
-            .EndRpc();
-
-        pc.SetLevel(level);
-        sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetLevel)
-            .Write(level)
-            .EndRpc();
-
-        sender.SendMessage();
-
-        DoppelPresentSkin[pc.PlayerId] = newOutfit;
-    }
-
     public override bool OnCheckMurderAsKiller(PlayerControl killer, PlayerControl target)
     {
         if (killer == null || target == null || Camouflage.IsCamouflage || Camouflager.AbilityActivated || Utils.IsActive(SystemTypes.MushroomMixupSabotage)) return true;
@@ -135,13 +78,13 @@ internal class Doppelganger : RoleBase
 
         DoppelVictim[target.PlayerId] = tname;
 
-        RpcChangeSkin(target, killerSkin, killerLvl);
+        target.SetNewOutfit(killerSkin, newLevel: killerLvl);
+        DoppelPresentSkin[target.PlayerId] = killerSkin;
         Logger.Info("Changed target skin", "Doppelganger");
-        RpcChangeSkin(killer, targetSkin, targetLvl);
-        Logger.Info("Changed killer skin", "Doppelganger");
 
-        //killer.Data.UpdateName(tname, Utils.GetClientById(killer.PlayerId));
-        //target.Data.UpdateName(kname, Utils.GetClientById(target.PlayerId));
+        killer.SetNewOutfit(targetSkin, newLevel: targetLvl);
+        DoppelPresentSkin[killer.PlayerId] = targetSkin;
+        Logger.Info("Changed killer skin", "Doppelganger");
 
         SendSkillRPC();
         RPC.SyncAllPlayerNames();
